Extract SMSCampaignForm
