+++ front/app/containers/Admin/messaging/texting/ExistingSMSCampaign/index.tsx
@@ -1,14 +1,11 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 
 // components
-import TextArea from 'components/UI/TextArea';
-import Button from 'components/UI/Button';
-import { Label, Box, IconTooltip } from '@citizenlab/cl2-component-library';
+import { Box } from '@citizenlab/cl2-component-library';
 import HelmetIntl from 'components/HelmetIntl';
 import TextingHeader from '../components/TextingHeader';
 import FormattedStatusLabel from '../components/FormattedStatusLabel';
-import RemainingCharacters from '../components/RemainingCharacters';
-
+import SMSCampaignForm from '../components/SMSCampaignForm';
 // i18n
 import { FormattedDate } from 'react-intl';
 
@@ -21,23 +18,21 @@ import useTextingCampaign from 'hooks/useTextingCampaign';
 
 // services
 import {
-  updateTextingCampaign,
   ITextingCampaignData,
-  ITextingCampaignStatuses,
+  TTextingCampaignStatus,
 } from 'services/textingCampaigns';
 
 // styling
 import styled from 'styled-components';
 import { isNilOrError } from 'utils/helperUtils';
 
-const StyledForm = styled.form`
+const StyledSMSCampaignForm = styled(SMSCampaignForm)`
   width: 500px;
 `;
 
-// enough to fit 3 messages, actual functionality TBD in subsequent ticket
-const MAX_CHAR_COUNT = 480;
-
-const getTitleMessage = (campaignStatus: ITextingCampaignStatuses) => {
+const getTitleMessage = (
+  campaignStatus: TTextingCampaignStatus | undefined
+) => {
   switch (campaignStatus) {
     case 'draft':
       return 'Draft SMS Campaign';
@@ -91,61 +86,11 @@ const ExistingSMSCampaign = (props: WithRouterProps) => {
   const { campaignId } = props.params;
   const campaign = useTextingCampaign(campaignId);
 
-  const [inputPhoneNumbers, setInputPhoneNumbers] = useState<string | null>(
-    null
-  );
-  const [inputMessage, setInputMessage] = useState<string | null>(null);
-  const [remainingChars, setRemainingChars] = useState(MAX_CHAR_COUNT);
-  const [hasPhoneNumbers, setHasPhoneNumbers] = useState(false);
-
-  const handleInputPhoneNumbersChange = (value: string) => {
-    setInputPhoneNumbers(value);
-    setHasPhoneNumbers(value.length > 0);
-  };
-
-  const handleInputMessageChange = (value: string) => {
-    setInputMessage(value);
-  };
-
-  const handleOnSubmit = async (event: React.FormEvent) => {
-    event.preventDefault();
-
-    if (isNilOrError(inputPhoneNumbers) || isNilOrError(inputMessage)) return;
-
-    const splitNumbers = inputPhoneNumbers.split(',');
-    try {
-      const result = await updateTextingCampaign(campaignId, {
-        message: inputMessage,
-        phone_numbers: splitNumbers,
-      });
-      const { id } = result.data;
-      const url = `/admin/messaging/texting/${id}/preview`;
-      clHistory.replace(url);
-    } catch (error) {
-      // handle error here in subsequent ticket
-    }
-  };
-
-  useEffect(() => {
-    if (!isNilOrError(campaign)) {
-      setInputMessage(campaign.attributes.message);
-      setInputPhoneNumbers(campaign.attributes.phone_numbers.join(', '));
-    }
-  }, [campaign]);
-
-  useEffect(() => {
-    if (isNilOrError(inputMessage)) return;
-    const remainingCharCount = MAX_CHAR_COUNT - inputMessage.length;
-    setRemainingChars(remainingCharCount);
-  }, [inputMessage]);
-
   // show campaign not found
   if (isNilOrError(campaign)) return null;
 
   const { status } = campaign.attributes;
   const isDraft = status === 'draft';
-  const overCharacterLimit = remainingChars < 0;
-  const isButtonDisabled = overCharacterLimit || hasPhoneNumbers;
 
   return (
     <>
@@ -166,57 +111,7 @@ const ExistingSMSCampaign = (props: WithRouterProps) => {
         </Box>
         <span>{getAdditionalInfoByStatus(campaign)}</span>
       </>
-      <StyledForm onSubmit={handleOnSubmit}>
-        <Box marginBottom="20px">
-          <Label>
-            {isDraft && (
-              <span>
-                Enter a list of phone numbers. Separate each number by a comma
-                and include the international dialing code (eg. +1).
-              </span>
-            )}
-            {!isDraft && <span>Sent to:</span>}
-          </Label>
-          <TextArea
-            rows={8}
-            maxRows={8}
-            value={inputPhoneNumbers}
-            disabled={!isDraft}
-            onChange={handleInputPhoneNumbersChange}
-          />
-        </Box>
-        <Box marginBottom="30px">
-          <Label>
-            Message <IconTooltip content="Help goes here" />
-          </Label>
-          <TextArea
-            rows={8}
-            maxRows={8}
-            value={inputMessage}
-            disabled={!isDraft}
-            onChange={handleInputMessageChange}
-          />
-          {isDraft && (
-            <RemainingCharacters
-              remainingChars={remainingChars}
-              overCharacterLimit={overCharacterLimit}
-            />
-          )}
-        </Box>
-
-        {isDraft && (
-          <Box display="flex" justifyContent="flex-start">
-            <Button
-              buttonStyle="primary"
-              size="2"
-              type="submit"
-              text={'Update and Preview SMS'}
-              onClick={handleOnSubmit}
-              disabled={isButtonDisabled}
-            />
-          </Box>
-        )}
-      </StyledForm>
+      <StyledSMSCampaignForm formIsLocked={!isDraft} campaign={campaign} />
     </>
   );
 };
